requirements.txt
fastapi
uvicorn
openai
google-analytics-data
pandas
pydantic
python-dotenv
requests
tabulate